+++ retail/swagger.py
@@ -1,0 +1,14 @@
+from drf_yasg.views import get_schema_view
+from drf_yasg import openapi
+from rest_framework import permissions
+
+
+view = get_schema_view(
+    openapi.Info(
+        title="Gallery API Documentation",
+        default_version="v3.2.14",
+        description="Documentation of the Gallery APIs",
+    ),
+    public=True,
+    permission_classes=(permissions.AllowAny,),
+).with_ui("swagger")